remove rpiinterface
