+++ wroprg/src/hardware/__init__.py
@@ -1,10 +1,12 @@
-"""This module initializes the hardware package by importing necessary classes"""
+"""Initialize the hardware package by exposing key classes."""
 from .hardware_interface import HardwareInterface
 from .validator import RobotValidator
 from .robotstate import RobotState
 from .orientation import OrientationEstimator
-from .rpi_interface import RpiInterface
 
-# We want to expose the HardwareInterface and RobotValidator classes
-# to be used in other modules, so we define them in __all__.
-__all__ = ["HardwareInterface", "RobotValidator", "RobotState","OrientationEstimator","RpiInterface"]
+__all__ = [
+	"HardwareInterface",
+	"RobotValidator",
+	"RobotState",
+	"OrientationEstimator",
+]
