+++ wroprg/src/hardware/hardware_interface.py
@@ -1,75 +1,243 @@
-"""Unified Hardware Interface for WRO Future Engineer 2025 project."""
+"""Unified Hardware Interface for WRO Future Engineer 2025 project.
+
+This merges the previous Raspberry Pi interface into a single class so
+all hardware access (LEGO + Raspberry Pi peripherals) is exposed here.
+"""
 
 import logging
-from typing import List, Optional
+import time
+import math
+from typing import List, Optional, Tuple
+from board import SCL, SDA
+import busio
+import adafruit_ssd1306
+import adafruit_mpu6050
+import adafruit_tca9548a
+import adafruit_vl53l0x
+from qmc5883l import QMC5883L
+from gpiozero import Buzzer, RGBLED, DistanceSensor, Button, Device
+from gpiozero.pins.pigpio import PiGPIOFactory
+from PIL import Image, ImageDraw, ImageFont
+
 from base.shutdown_handling import ShutdownInterface
 from hardware.robotstate import RobotState
 from hardware.legodriver import BuildHatDriveBase
 from hardware.measurements import MeasurementFileLog
 from hardware.orientation import OrientationEstimator
-from hardware.rpi_interface import RpiInterface
 from hardware.camerameasurements import CameraDistanceMeasurements
+from hardware.screenlogger import ScreenLogger
+from hardware.camera import MyCamera
+from utils import constants
 
 logger = logging.getLogger(__name__)
+
 
 class HardwareInterface(ShutdownInterface):
     """
     Provides unified access to all hardware components.
-    Includes methods for both LEGO driver and Raspberry Pi interface.
+    Includes methods for both LEGO driver and Raspberry Pi peripherals.
     """
 
-    def __init__(self,stabilize:bool) -> None:
+    # --- Constants and defaults (from former RpiInterface) ---
+    MAX_STABILIZATION_CHECKS = 5
+    LINE_HEIGHT = 10  # pixels per line
+    FONT_SIZE = 10  # font size for messages
+    LEFT_LASER_CHANNEL = 7
+    RIGHT_LASER_CHANNEL = 2
+    DEVICE_I2C_CHANNEL = 6
+    DISTANCE_FUSION = True
 
+    DISTANCE_SENSOR_DISTANCE = 12.5  # cm distance between sensors
+
+    _screenlogger: Optional[ScreenLogger] = None
+    display_loglines = True
+
+    BUZZER_PIN = 13
+    LED1_RED_PIN = 12
+    LED1_GREEN_PIN = 6
+    LED1_BLUE_PIN = 5
+    BUTTON_PIN = 19
+    RIGHT_SENSOR_TRIG_PIN = 23
+    RIGHT_SENSOR_ECHO_PIN = 21
+    RIGHT_DISTANCE_MAX_DISTANCE = 2
+    LEFT_SENSOR_TRIG_PIN = 20
+    LEFT_SENSOR_ECHO_PIN = 24
+    LEFT_DISTANCE_MAX_DISTANCE = 2
+    FRONT_SENSOR_TRIG_PIN = 27
+    FRONT_SENSOR_ECHO_PIN = 22
+    FRONT_DISTANCE_MAX_DISTANCE = 2
+    JUMPER_PIN = 26
+
+    # Screen settings
+    SCREEN_WIDTH = 128
+    SCREEN_HEIGHT = 64
+    SCREEN_UPDATE_INTERVAL = 0.5  # seconds
+    LED_TEST_DELAY = 0.05  # seconds
+
+    def __init__(self, stabilize: bool) -> None:
+        # LEGO drive base initialization
         self._lego_drive_base: Optional[BuildHatDriveBase] = None
         self._full_initialization()
-        self._rpi = RpiInterface(stabilize)
 
-        self._measurements_manager: Optional[MeasurementFileLog] = None
+        # Raspberry Pi peripherals initialization (merged from RpiInterface)
+        logger.info("Initializing Raspberry Pi peripherals...")
 
-        self._measurements_manager = MeasurementFileLog(self)
+        # Setup I2C devices
+        i2c = busio.I2C(SCL, SDA)
+        tca = adafruit_tca9548a.TCA9548A(i2c)
 
-        self._orientation_estimator = OrientationEstimator(
-                get_accel=self.get_acceleration,
-                get_gyro=self.get_gyro,
-                get_mag=self.get_magnetometer
+        for channel in range(8):
+            if tca[channel].try_lock():
+                logger.info("Channel %s:", channel)
+                addresses = tca[channel].scan()
+                logger.info([hex(address) for address in addresses if address != 0x70])
+                tca[channel].unlock()
+
+        device_channel = tca[self.DEVICE_I2C_CHANNEL]
+
+        # Sensors on I2C
+        self.mpu = adafruit_mpu6050.MPU6050(device_channel)
+        self.compass = QMC5883L(device_channel)
+
+        # OLED display
+        self.oled = adafruit_ssd1306.SSD1306_I2C(self.SCREEN_WIDTH, self.SCREEN_HEIGHT, device_channel)
+        self.oled.fill(0)
+        self.oled.show()
+
+        self.font = ImageFont.truetype(
+            "/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf", HardwareInterface.FONT_SIZE
+        )
+        self._last_oled_update: float = 0
+        self.image: Image.Image = Image.new("1", (self.SCREEN_WIDTH, self.SCREEN_HEIGHT))
+        self.draw: ImageDraw.ImageDraw = ImageDraw.Draw(self.image)
+
+        self.pendingmessage: bool = False
+        self.messages: List[str] = []
+
+        # Laser distance sensors via TCA channels
+        left_channel = tca[self.LEFT_LASER_CHANNEL]
+        right_channel = tca[self.RIGHT_LASER_CHANNEL]
+        self.left_laser = adafruit_vl53l0x.VL53L0X(left_channel)
+        self.left_laser.measurement_timing_budget = 200000
+        self.right_laser = adafruit_vl53l0x.VL53L0X(right_channel)
+        self.right_laser.measurement_timing_budget = 200000
+
+        self.display_message("Initializing Pi Interface...")
+
+        # GPIO via pigpio if available
+        try:
+            Device.pin_factory = PiGPIOFactory()
+            logger.info("Using PiGPIOFactory for GPIO pin control.")
+        except Exception:  # pylint: disable=broad-except
+            Device.pin_factory = None
+            logger.error("Failed to initialize PiGPIOFactory")
+
+        # Actuators and buttons
+        self.buzzer = Buzzer(self.BUZZER_PIN)
+        self.led1 = RGBLED(
+            red=self.LED1_RED_PIN,
+            green=self.LED1_GREEN_PIN,
+            blue=self.LED1_BLUE_PIN,
         )
 
-        self.camera_measurements = CameraDistanceMeasurements(
-                                    self._rpi.get_camera())
+        # LED test
+        self.led1.color = (0, 1, 0)
+        time.sleep(self.LED_TEST_DELAY)
+        self.led1.color = (1, 0, 0)
+        time.sleep(self.LED_TEST_DELAY)
+        self.led1.color = (0, 0, 1)
+        time.sleep(self.LED_TEST_DELAY)
+        self.led1.color = (0, 0, 0)
 
-    def clear_messages(self) -> None:
-        """clear display messages"""
-        if self._rpi is None:
-            raise RuntimeError("Raspberry Pi interface not initialized." \
-            " Call full_initialization() first.")
-        else:
-            return self._rpi.clear_messages()
+        self.action_button = Button(self.BUTTON_PIN, hold_time=1)
+
+        # Ultrasonic distance sensors
+        self.rightdistancesensor = DistanceSensor(
+            echo=self.RIGHT_SENSOR_ECHO_PIN,
+            trigger=self.RIGHT_SENSOR_TRIG_PIN,
+            partial=True,
+            max_distance=self.RIGHT_DISTANCE_MAX_DISTANCE,
+        )
+        self.leftdistancesensor = DistanceSensor(
+            echo=self.LEFT_SENSOR_ECHO_PIN,
+            trigger=self.LEFT_SENSOR_TRIG_PIN,
+            partial=True,
+            max_distance=self.LEFT_DISTANCE_MAX_DISTANCE,
+        )
+        self.front_distance_sensor = DistanceSensor(
+            echo=self.FRONT_SENSOR_ECHO_PIN,
+            trigger=self.FRONT_SENSOR_TRIG_PIN,
+            partial=True,
+            max_distance=self.FRONT_DISTANCE_MAX_DISTANCE,
+        )
+        self.jumper_pin = Button(self.JUMPER_PIN, hold_time=1)
+
+        # Camera
+        self.camera = MyCamera()
+
+        logger.info("Raspberry Pi peripherals initialized successfully.")
+
+        if stabilize:
+            logger.warning("Stabilize Distance Sensors...")
+            time.sleep(0.5)
+            counter = 0
+            valid_distance = False
+            while counter < self.MAX_STABILIZATION_CHECKS and not valid_distance:
+                valid_distance = True
+                if (self.get_right_distance() < 0.1 or
+                        self.get_right_distance() >= constants.RIGHT_DISTANCE_MAX):
+                    logger.info("Right distance sensor is not stable, %.2f cm", self.get_right_distance())
+                    valid_distance = False
+                if (self.get_left_distance() < 0.1 or
+                        self.get_left_distance() >= constants.LEFT_DISTANCE_MAX):
+                    logger.info("Left distance sensor is not stable, %.2f cm", self.get_left_distance())
+                    valid_distance = False
+                if not valid_distance:
+                    logger.warning("Waiting for distance sensors to stabilize...")
+                    time.sleep(1)
+                counter += 1
+
+        # Measurements and orientation
+        self._measurements_manager: Optional[MeasurementFileLog] = MeasurementFileLog(self)
+        self._orientation_estimator = OrientationEstimator(
+            get_accel=self.get_acceleration,
+            get_gyro=self.get_gyro,
+            get_mag=self.get_magnetometer,
+        )
+        self.camera_measurements = CameraDistanceMeasurements(self.get_camera())
 
     def _full_initialization(self) -> None:
         """Initialize all hardware components."""
-        self._lego_drive_base = BuildHatDriveBase(front_motor_port='D', back_motor_port='B',
-                                               bottom_color_sensor_port='C')
+        self._lego_drive_base = BuildHatDriveBase(
+            front_motor_port='D', back_motor_port='B', bottom_color_sensor_port='C'
+        )
 
-    def wait_for_ready(self):
+    def wait_for_ready(self) -> None:
         """Wait for complete hardware initialization."""
         if self._lego_drive_base is not None:
             self._lego_drive_base.wait_for_setup()
 
-
+    # --- LiDAR distances ---
     def get_right_lidar_distance(self) -> float:
         """Get the distance from the right lidar sensor."""
-        return self._rpi.get_right_lidar_distance()
+        return self.right_laser.range / 10.0
 
     def get_left_lidar_distance(self) -> float:
         """Get the distance from the left lidar sensor"""
-        return self._rpi.get_left_lidar_distance()
+        return self.left_laser.range / 10.0
 
+    # --- Fused wall distances (public wrappers) ---
+    def get_right_distance(self) -> float:
+        return self._get_right_distance()
+
+    def get_left_distance(self) -> float:
+        return self._get_left_distance()
+
+    # --- Measurements and orientation management ---
     def start_measurement_recording(self) -> None:
         """Start the measurements manager thread."""
         if self._measurements_manager is None:
-            raise RuntimeError("Measurements manager not initialized. Call" \
-                    " full_initialization() first.")
-
+            raise RuntimeError("Measurements manager not initialized. Call full_initialization() first.")
         if self._orientation_estimator is not None:
             self._orientation_estimator.start_readings()
         self._measurements_manager.start_reading()
@@ -80,16 +248,16 @@ class HardwareInterface(ShutdownInterface):
         if self._measurements_manager is not None:
             self._measurements_manager.add_comment(comment)
 
-    # --- Raspberry Pi Interface Methods ---
-
+    # --- Raspberry Pi Peripheral Methods ---
     def log_message(self, front: float, left: float, right: float, current_yaw: float,
-                                                            current_steering: float) -> None:
+                    current_steering: float) -> None:
         """Log the sensor readings and robot state."""
-        if self._rpi is None:
-            raise RuntimeError("Raspberry Pi interface not initialized.")
-        self._rpi.log_message(front, left, right, current_yaw, current_steering)
+        image: Image.Image = self.get_screen_logger().log_message(
+            front, left, right, current_yaw, current_steering
+        )
+        self.paint_display(image)
 
-    def get_orientation(self):
+    def get_orientation(self) -> Tuple[float, float, float]:
         """Get the current (roll, pitch, yaw) in degrees."""
         if self._orientation_estimator is None:
             raise RuntimeError("Orientation estimator not initialized.")
@@ -97,59 +265,80 @@ class HardwareInterface(ShutdownInterface):
 
     def buzzer_beep(self, timer: float = 0.5) -> None:
         """Turn on the buzzer."""
-        self._rpi.buzzer_beep(timer)
+        self.buzzer.on()
+        time.sleep(timer)
+        self.buzzer.off()
 
     def led1_green(self) -> None:
         """Turn on the LED1 green."""
-        self._rpi.led1_green()
+        self.led1.color = (0, 1, 0)
 
     def led1_red(self) -> None:
         """Turn on the LED1 red."""
-        self._rpi.led1_red()
+        self.led1.color = (1, 0, 0)
 
     def led1_blue(self) -> None:
         """Turn on the LED1 blue."""
-        self._rpi.led1_blue()
+        self.led1.color = (0, 0, 1)
 
     def led1_white(self) -> None:
         """Turn on the LED1 white."""
-        self._rpi.led1_white()
+        self.led1.color = (1, 1, 1)
 
     def led1_off(self) -> None:
         """Turn off the LED1."""
-        self._rpi.led1_off()
+        self.led1.color = (0, 0, 0)
 
     def wait_for_action(self) -> None:
         """Wait for the action button to be pressed."""
-        self._rpi.wait_for_action()
+        logger.warning("Waiting for action button press...")
+        self.action_button.wait_for_active()
+        logger.info("Action button pressed!")
 
+    # --- Distance helpers ---
     def _get_right_distance(self) -> float:
         """Get the distance from the right distance sensor."""
-        return self._rpi.get_right_distance()
+        ultrasonic = self.rightdistancesensor.distance * 100  # cm
+        laser = self.right_laser.range / 10.0  # cm
+        return self._fuse_sensors(laser, ultrasonic)
 
     def _get_left_distance(self) -> float:
         """Get the distance from the left distance sensor."""
-        return self._rpi.get_left_distance()
+        ultrasonic = self.leftdistancesensor.distance * 100  # cm
+        laser = self.left_laser.range / 10.0  # cm
+        return self._fuse_sensors(laser, ultrasonic)
 
+    def _get_front_distance(self) -> float:
+        """Get the distance to the front obstacle in centimeter."""
+        return self.front_distance_sensor.distance * 100
+
+    # --- Display helpers ---
     def display_message(self, message: str, forceflush: bool = False) -> None:
         """
         Display a message on the OLED screen.
 
         Only the last 5 messages are shown on the display.
         """
-        self._rpi.display_message(message, forceflush)
+        now = time.time()
+        self.messages.append(message)
+        self.messages = self.messages[-5:]
+        if forceflush or (now - self._last_oled_update >= self.SCREEN_UPDATE_INTERVAL):
+            self.flush_pending_messages()
+            self.pendingmessage = False
 
     def force_flush_messages(self) -> None:
         """Force flush the messages on the OLED screen."""
-        self._rpi.force_flush_messages()
+        if self.pendingmessage:
+            self.flush_pending_messages()
+            self.pendingmessage = False
 
     def add_screen_logger_message(self, message: List[str]) -> None:
         """Add a message to the screen logger."""
-        self._rpi.add_screen_logger_message(message)
+        self.get_screen_logger().add_message(message)
 
     def get_jumper_state(self) -> bool:
-        """Get the state of the jumper pin."""        
-        return self._rpi.get_jumper_state()
+        """Get the state of the jumper pin."""
+        return self.jumper_pin.is_active
 
     def shutdown(self) -> None:
         """Shutdown the hardware interface."""
@@ -159,7 +348,48 @@ class HardwareInterface(ShutdownInterface):
             self._lego_drive_base.shutdown()
         self.camera_measurements.shutdown()
 
-        self._rpi.shutdown()
+        # Shutdown Raspberry Pi peripherals
+        try:
+            self.flush_pending_messages()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error flushing OLED messages during shutdown: %s", e)
+        try:
+            if self.camera is not None:
+                self.camera.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing camera during shutdown: %s", e)
+        try:
+            self.buzzer.off()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error turning off buzzer during shutdown: %s", e)
+        try:
+            self.led1.color = (0, 0, 0)
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error turning off LED1 during shutdown: %s", e)
+        try:
+            self.led1.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing LED1 during shutdown: %s", e)
+        try:
+            self.buzzer.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing buzzer during shutdown: %s", e)
+        try:
+            self.rightdistancesensor.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing right distance sensor during shutdown: %s", e)
+        try:
+            self.leftdistancesensor.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing left distance sensor during shutdown: %s", e)
+        try:
+            self.front_distance_sensor.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing front distance sensor during shutdown: %s", e)
+        try:
+            self.jumper_pin.close()
+        except Exception as e:  # pylint: disable=broad-except
+            logger.error("Error closing jumper pin during shutdown: %s", e)
         if self._orientation_estimator is not None:
             self._orientation_estimator.shutdown()
         if self._measurements_manager is not None:
@@ -171,36 +401,31 @@ class HardwareInterface(ShutdownInterface):
             raise RuntimeError("LEGO Drive Base not initialized. Call full_initialization() first.")
         self._lego_drive_base.reset_front_motor()
 
-    def get_acceleration(self) -> tuple[float, float, float]:
+    def get_acceleration(self) -> Tuple[float, float, float]:
         """Get the acceleration from the MPU6050 sensor."""
-        return self._rpi.get_acceleration()
+        accel = self.mpu.acceleration
+        return accel
 
-    def get_magnetometer(self) -> tuple[float, float, float]:
+    def get_magnetometer(self) -> Tuple[float, float, float]:
         """Get the magnetometer data from the QMC5883L sensor."""
-        return self._rpi.get_magnetometer()
+        return self.compass.magnetic
 
-    def get_gyro(self) -> tuple[float, float, float]:
+    def get_gyro(self) -> Tuple[float, float, float]:
         """Get the gyroscope data from the MPU6050 sensor."""
-        return self._rpi.get_gyro()
+        return self.mpu.gyro
 
-    def reset_gyro(self)-> None:
+    def reset_gyro(self) -> None:
         """Reset the yaw angle to zero."""
         if self._orientation_estimator is not None:
-            # self._orientation_estimator.reset()
             self._orientation_estimator.reset_yaw()
         else:
             raise RuntimeError("Orientation estimator not initialized.")
 
-    def is_button_pressed(self):
+    def is_button_pressed(self) -> bool:
         """Check if the action button is pressed."""
-        if self._rpi is None:
-            raise RuntimeError("Raspberry Pi interface not initialized.\
-                                Call full_initialization() first.")
-        else:
-            return self._rpi.action_button.is_active
+        return self.action_button.is_active
 
     # --- LEGO Driver Methods ---
-
     def camera_off(self) -> None:
         """Turn off the camera."""
         if self._lego_drive_base is None:
@@ -225,7 +450,7 @@ class HardwareInterface(ShutdownInterface):
             raise RuntimeError("LEGO Drive Base not initialized. Call full_initialization() first.")
         self._lego_drive_base.run_front(-speed)
 
-    def turn_steering(self, degrees: float, steering_speed: float=40) -> None:
+    def turn_steering(self, degrees: float, steering_speed: float = 40) -> None:
         """
         Turn the steering by the specified degrees.
         Positive degrees turn right, negative turn left.
@@ -253,10 +478,6 @@ class HardwareInterface(ShutdownInterface):
             raise RuntimeError("LEGO Drive Base not initialized. Call full_initialization() first.")
         return self._lego_drive_base.get_bottom_color_rgbi()
 
-    def _get_front_distance(self) -> float:
-        """Get the distance to the front obstacle in centimeter."""
-        return self._rpi.get_front_distance()
-
     def get_steering_angle(self) -> float:
         """Get the current steering angle in degrees."""
         if self._lego_drive_base is None:
@@ -272,12 +493,93 @@ class HardwareInterface(ShutdownInterface):
         yaw = self.get_orientation()[2]
 
         (camera_front, camera_left, camera_right, _) = self.camera_measurements.get_distance()
-        return RobotState(front=front, left=left, right=right, yaw=yaw,
-                        camera_front=camera_front,
-                        camera_left=camera_left,
-                        camera_right=camera_right)
-        # return RobotState(front=front, left=left, right=right, yaw=yaw)
+        return RobotState(
+            front=front,
+            left=left,
+            right=right,
+            yaw=yaw,
+            camera_front=camera_front,
+            camera_left=camera_left,
+            camera_right=camera_right,
+        )
 
     def disable_logger(self) -> None:
         """Disable the logger."""
-        self._rpi.disable_logger()
+        self.display_loglines = False
+
+    # -------------------- Additional helpers from former RpiInterface --------------------
+    def get_camera(self) -> MyCamera:
+        return self.camera
+
+    def get_screen_logger(self) -> ScreenLogger:
+        if self._screenlogger is None:
+            self._screenlogger = ScreenLogger(width=self.SCREEN_WIDTH, height=self.SCREEN_HEIGHT)
+        return self._screenlogger
+
+    def paint_display(self, img: Image.Image) -> None:
+        self.oled.image(img)
+        self.oled.show()
+
+    def clear_messages(self) -> None:
+        self.messages.clear()
+
+    def flush_pending_messages(self) -> None:
+        now = time.time()
+        if self.display_loglines:
+            self.draw.rectangle((0, 0, self.SCREEN_WIDTH, self.SCREEN_HEIGHT), outline=0, fill=0)
+            for i, msg in enumerate(self.messages):
+                self.draw.text((0, i * HardwareInterface.LINE_HEIGHT), msg, font=self.font, fill=255)
+            self.oled.image(self.image)
+            self.oled.show()
+        self._last_oled_update = now
+
+    def enable_logger(self) -> None:
+        self.display_loglines = True
+
+    def get_front_distance_max(self) -> float:
+        return self.FRONT_DISTANCE_MAX_DISTANCE * 100
+
+    def _fuse_sensors(
+        self,
+        lidar_val: float,
+        ultrasonic_val: float,
+        lidar_weight: float = 0.7,
+        ultrasonic_weight: float = 0.3,
+        max_diff: float = 20,
+    ) -> float:
+        if lidar_val > 0 and ultrasonic_val > 0:
+            if abs(lidar_val - ultrasonic_val) > max_diff:
+                return lidar_val
+            fused = (lidar_val * lidar_weight + ultrasonic_val * ultrasonic_weight) / (
+                lidar_weight + ultrasonic_weight
+            )
+            return fused
+        return lidar_val if lidar_val > 0 else ultrasonic_val
+
+    def get_left_wangle(self) -> float:
+        ultrasonic = self.leftdistancesensor.distance * 100
+        laser = self.left_laser.range / 10.0
+        return self._wall_angle(front_dist=laser, back_dist=ultrasonic)
+
+    def get_right_wangle(self) -> float:
+        ultrasonic = self.rightdistancesensor.distance * 100
+        laser = self.right_laser.range / 10.0
+        return self._wall_angle(front_dist=laser, back_dist=ultrasonic)
+
+    def get_left_pdistance(self) -> float:
+        ultrasonic = self.leftdistancesensor.distance * 100
+        laser = self.left_laser.range / 10.0
+        return self._perpendicular_distance(front_dist=laser, back_dist=ultrasonic)
+
+    def get_right_pdistance(self) -> float:
+        ultrasonic = self.rightdistancesensor.distance * 100
+        laser = self.right_laser.range / 10.0
+        return self._perpendicular_distance(front_dist=laser, back_dist=ultrasonic)
+
+    def _wall_angle(self, front_dist: float, back_dist: float, sensor_gap: float = DISTANCE_SENSOR_DISTANCE) -> float:
+        theta_rad = math.atan2(front_dist - back_dist, sensor_gap)
+        theta_deg = math.degrees(theta_rad)
+        return theta_deg
+
+    def _perpendicular_distance(self, front_dist: float, back_dist: float) -> float:
+        return (front_dist + back_dist) / 2.0
