--- wroprg/src/hardware/rpi_interface.py
@@ -1,467 +1,0 @@
-"""This module is to interact with all hardware connected to RasperryPi."""
-import logging
-import time
-from typing import List,Optional, Tuple
-import math
-from board import SCL, SDA
-import busio
-import adafruit_ssd1306
-import adafruit_mpu6050
-import adafruit_tca9548a
-from qmc5883l import QMC5883L
-import adafruit_vl53l0x
-from gpiozero import Buzzer, RGBLED, DistanceSensor, Button, Device
-from gpiozero.pins.pigpio import PiGPIOFactory
-from PIL import Image, ImageDraw, ImageFont
-from hardware.screenlogger import ScreenLogger
-from hardware.camera import MyCamera
-from utils import constants
-from base.shutdown_handling import ShutdownInterface
-
-logger: logging.Logger = logging.getLogger(__name__)
-class RpiInterface(ShutdownInterface):
-    """ This interface defines all Interfaces on Raspberry Pi."""
-
-    # Use the centralized pin configuration class
-
-    MAX_STABILIZATION_CHECKS = 5
-    LINE_HEIGHT = 10  # pixels per line
-    FONT_SIZE = 10 # font size for messages
-    LEFT_LASER_CHANNEL = 7
-    RIGHT_LASER_CHANNEL = 2
-    DEVICE_I2C_CHANNEL=6
-    DISTANCE_FUSION = True
-
-    DISTANCE_SENSOR_DISTANCE = 12.5  # distance between sensors in cm
-
-
-    _screenlogger: Optional[ScreenLogger] = None
-    display_loglines = True
-
-    BUZZER_PIN = 13
-    LED1_RED_PIN = 12
-    LED1_GREEN_PIN = 6
-    LED1_BLUE_PIN =  5
-    BUTTON_PIN =  19
-    RIGHT_SENSOR_TRIG_PIN = 23
-    RIGHT_SENSOR_ECHO_PIN = 21
-    RIGHT_DISTANCE_MAX_DISTANCE = 2
-    LEFT_SENSOR_TRIG_PIN =  20
-    LEFT_SENSOR_ECHO_PIN = 24
-    LEFT_DISTANCE_MAX_DISTANCE =  2
-    FRONT_SENSOR_TRIG_PIN = 27
-    FRONT_SENSOR_ECHO_PIN = 22
-    FRONT_DISTANCE_MAX_DISTANCE = 2
-    JUMPER_PIN =  26
-
-
-    # Common settings for all chassis versions
-    SCREEN_WIDTH = 128
-    SCREEN_HEIGHT = 64
-    SCREEN_UPDATE_INTERVAL = 0.5  # seconds
-    LED_TEST_DELAY = 0.05  # seconds
-
-
-    def __init__(self,stabilize:bool) -> None:
-        """
-        Initialize the LED control class.
-        Sets up the GPIO pins for the LEDs and initializes the buzzer and RGB LED.
-        """
-        super().__init__()
-
-        logger.info("Initializing RpiInterface...")
-
-        #Setup I2c devices.
-        i2c = busio.I2C(SCL,SDA)  # uses board.SCL and board.SDA
-
-        tca = adafruit_tca9548a.TCA9548A(i2c)
-
-        for channel in range(8):
-            if tca[channel].try_lock():
-                logger.info("Channel {}:".format(channel))
-                addresses = tca[channel].scan()
-                logger.info([hex(address) for address in addresses if address != 0x70])
-                tca[channel].unlock()
-
-        device_channel= tca[self.DEVICE_I2C_CHANNEL]
-
-        # Initialize MPU6050 sensor
-        self.mpu = adafruit_mpu6050.MPU6050(device_channel)
-
-        self.compass = QMC5883L(device_channel)
-
-        # Create the SSD1306 OLED class.
-        self.oled = adafruit_ssd1306.SSD1306_I2C(self.SCREEN_WIDTH,
-                                                 self.SCREEN_HEIGHT, device_channel)
-
-        # Clear display.
-        self.oled.fill(0)
-        self.oled.show()
-
-        self.font = ImageFont.truetype("/usr/share/fonts/truetype/dejavu/DejaVuSans.ttf",
-                                        RpiInterface.FONT_SIZE)
-        # Load a default font
-        #self.font: ImageFont.FreeTypeFont | ImageFont.ImageFont = ImageFont.load_default()
-        self._last_oled_update: float = 0
-        self.image: Image.Image = Image.new("1", (self.SCREEN_WIDTH, self.SCREEN_HEIGHT))
-        self.draw: ImageDraw.ImageDraw = ImageDraw.Draw(self.image)
-
-        self.pendingmessage: bool = False  # Initialize pendingmessage flag
-        self.messages: List[str] = []  # Initialize messages list
-
-        #intialize laser distance sensor
-        left_channel = tca[self.LEFT_LASER_CHANNEL]
-        right_channel = tca[self.RIGHT_LASER_CHANNEL]
-
-        self.left_laser = adafruit_vl53l0x.VL53L0X(left_channel)
-        self.left_laser.measurement_timing_budget = 200000
-        self.right_laser = adafruit_vl53l0x.VL53L0X(right_channel)
-        self.right_laser.measurement_timing_budget = 200000
-
-        #Logger is not setup yet, so we use print for initialization messages
-        self.display_message("Initializing Pi Interface...")
-
-
-        try:
-            # Use pigpio factory if available
-            Device.pin_factory = PiGPIOFactory()
-            logger.info("Using PiGPIOFactory for GPIO pin control.")
-        except : # pylint: disable=bare-except
-            # Fallback to default GPIO pin factory
-            Device.pin_factory = None
-            logger.error("Failed to initialize PiGPIOFactory")
-
-        self.buzzer = Buzzer(self.BUZZER_PIN)
-        self.led1 = RGBLED(red=self.LED1_RED_PIN,
-                           green=self.LED1_GREEN_PIN,
-                           blue=self.LED1_BLUE_PIN)
-        """Turn on the LED1 Test."""
-        self.led1.color = (0, 1, 0)  # green
-        time.sleep(self.LED_TEST_DELAY)
-        self.led1.color = (1, 0, 0)  # red
-        time.sleep(self.LED_TEST_DELAY)
-        self.led1.color = (0, 0, 1)  # blue
-        time.sleep(self.LED_TEST_DELAY)
-        self.led1.color = (0, 0, 0)  # off
-
-        # Set up the button
-        self.action_button = Button(self.BUTTON_PIN, hold_time=1)
-
-        self.rightdistancesensor = DistanceSensor(echo=self.RIGHT_SENSOR_ECHO_PIN,
-                                                  trigger=self.RIGHT_SENSOR_TRIG_PIN,
-                                                  partial=True,
-                                                  max_distance=
-                                                  self.RIGHT_DISTANCE_MAX_DISTANCE)
-        self.leftdistancesensor = DistanceSensor(echo=self.LEFT_SENSOR_ECHO_PIN,
-                                                 trigger=self.LEFT_SENSOR_TRIG_PIN,
-                                                 partial=True,
-                                                 max_distance=self.LEFT_DISTANCE_MAX_DISTANCE)
-
-        # Initialize Front Distance Sensor
-        self.front_distance_sensor = DistanceSensor(echo=self.FRONT_SENSOR_ECHO_PIN,
-                                                        trigger=self.FRONT_SENSOR_TRIG_PIN,
-                                                        partial=True,
-                                                        max_distance=
-                                                        self.FRONT_DISTANCE_MAX_DISTANCE)
-        self.jumper_pin = Button(self.JUMPER_PIN, hold_time=1)
-
-
-        #setup camera library
-        self.camera = MyCamera()
-
-        logger.info("RpiInterface initialized successfully.")
-
-        if stabilize:
-            logger.warning("Stabilize Distance Sensors...")
-            # Stabilize distance sensors
-            time.sleep(0.5)  # Wait for sensors to stabilize
-            counter = 0
-            valid_distance = False
-
-            while counter< self.MAX_STABILIZATION_CHECKS and valid_distance is False:
-                valid_distance = True
-                if (self.get_right_distance() < 0.1 or
-                        self.get_right_distance() >= constants.RIGHT_DISTANCE_MAX):
-                    logger.info("Right distance sensor is not stable, %.2f cm",
-                                    self.get_right_distance())
-                    valid_distance = False
-                if (self.get_left_distance() < 0.1 or
-                        self.get_left_distance() >= constants.LEFT_DISTANCE_MAX):
-                    logger.info("Left distance sensor is not stable, %.2f cm",
-                                    self.get_left_distance())
-                    valid_distance = False
-                if valid_distance is False:
-                    logger.warning("Waiting for distance sensors to stabilize...")
-                    time.sleep(1)
-                counter += 1
-
-    def get_magnetometer(self) -> Tuple[float, float, float]:
-        """Get the current magnetometer reading (uT) as (mx, my, mz)."""
-        return self.compass.magnetic
-
-    def get_camera(self) -> MyCamera:
-        """Get the camera instance."""
-        return self.camera
-
-    def get_screen_logger(self) -> ScreenLogger:
-        """Get the logger for the RpiInterface."""
-        if self._screenlogger is None:
-            self._screenlogger = ScreenLogger(width=self.SCREEN_WIDTH,height=self.SCREEN_HEIGHT)
-        return self._screenlogger
-
-    def add_screen_logger_message(self, message: List[str]) -> None:
-        """Add a message to the screen logger."""
-        self.get_screen_logger().add_message(message)
-
-    def log_message(self, front: float, left: float, right: float, current_yaw: float,
-                                                            current_steering: float) -> None:
-        """Log a message to the screen."""
-        image: Image.Image = self.get_screen_logger().log_message(front, left, right,
-                                                                current_yaw, current_steering)
-        self.paint_display(image)
-
-
-    def disable_logger(self) -> None:
-        """Disable the logger."""
-        self.display_loglines = False
-
-    def enable_logger(self) -> None:
-        """Enable the logger."""
-        self.display_loglines = True
-
-    def buzzer_beep(self, timer: float = 0.5) -> None:
-        """Turn on the buzzer."""
-        self.buzzer.on()
-        time.sleep(timer)
-        self.buzzer.off()
-
-    def led1_green(self) -> None:
-        """Turn on the LED1 green."""
-        self.led1.color = (0, 1, 0)  # green
-
-    def led1_red(self) -> None:
-        """Turn on the LED1 red."""
-        self.led1.color = (1, 0, 0)  # red
-
-    def led1_blue(self) -> None:
-        """Turn on the LED1 blue."""
-        self.led1.color = (0, 0, 1)  # blue
-
-    def led1_white(self) -> None:
-        """Turn on the LED1 white."""
-        self.led1.color = (1, 1, 1)
-
-    def led1_off(self) -> None:
-        """Turn off the LED1."""
-        self.led1.color = (0, 0, 0)  # off
-
-    def wait_for_action(self) -> None:
-        """Wait for the action button to be pressed."""
-        logger.warning("Waiting for action button press...")
-        self.action_button.wait_for_active()
-        logger.info("Action button pressed!")
-
-    def _fuse_sensors(self,lidar_val, ultrasonic_val, lidar_weight=0.7, \
-                      ultrasonic_weight=0.3, max_diff=20):
-        """
-        Fuses LiDAR and ultrasonic readings using weighted average.
-        Rejects outliers if difference is too big.
-        """
-        if lidar_val > 0 and ultrasonic_val > 0:
-            if abs(lidar_val - ultrasonic_val) > max_diff:
-                # Trust LiDAR more if readings differ too much
-                return lidar_val
-            else:
-                # Weighted average
-                fused = (lidar_val * lidar_weight + ultrasonic_val * ultrasonic_weight) \
-                            / (lidar_weight + ultrasonic_weight)
-                return fused
-        else:
-            #if either is 0 return the other.
-            return lidar_val if lidar_val > 0 else ultrasonic_val
-
-
-    def get_right_lidar_distance(self) -> float:
-        """Get the distance from the right lidar sensor."""
-        return self.right_laser.range / 10.0  # Convert mm to cm
-
-    def get_left_lidar_distance(self) -> float:
-        """Get the distance from the left lidar sensor"""
-        return self.left_laser.range / 10.0  # Convert mm to cm
-
-    def get_right_distance(self) -> float:
-        """Get the distance from the distance sensor."""
-        ultrasonic = self.rightdistancesensor.distance * 100  # Convert to cm
-        laser = self.right_laser.range / 10.0  # Convert mm to cm
-        return self._fuse_sensors(laser, ultrasonic)
-
-    def _min_distance(self,ultrasonic:float,laser:float,sensor:str)->float:
-        if laser < ultrasonic and laser > 0:
-            # logger.info("Using Laser for laser:%.2f ultra:%.2f",laser,ultrasonic)
-            return laser
-        elif ultrasonic > 0:
-            return ultrasonic
-        elif laser > 0:
-            return laser
-        logger.info("No value for sensor %s",sensor)
-        return 0
-
-    def get_left_distance(self) -> float:
-        """Get the distance from the distance sensor."""
-
-        ultrasonic = self.leftdistancesensor.distance * 100  # Convert to cm
-        laser = self.left_laser.range / 10.0  # Convert mm to cm
-
-        return self._fuse_sensors(laser, ultrasonic)
-
-    def get_left_wangle(self) -> float:
-        """Get the wall angle from the left wall."""
-        ultrasonic = self.leftdistancesensor.distance * 100  # Convert to cm
-        laser = self.left_laser.range / 10.0  # Convert mm to cm
-        return self._wall_angle(front_dist=laser,
-                           back_dist=ultrasonic)
-
-    def get_right_wangle(self)->float:
-        """Get the wall angle from the right wall"""
-        ultrasonic = self.rightdistancesensor.distance * 100  # Convert to cm
-        laser = self.right_laser.range / 10.0  # Convert mm to cm
-        return self._wall_angle(front_dist=laser,
-                           back_dist=ultrasonic)
-
-    def get_left_pdistance(self)->float:
-        """Get the perpendicular distance from the left wall."""
-        ultrasonic = self.leftdistancesensor.distance * 100  # Convert to cm
-        laser = self.left_laser.range / 10.0  # Convert mm to cm
-        return self._perpendicular_distance(front_dist=laser,
-                                             back_dist=ultrasonic)
-
-    def get_right_pdistance(self)->float:
-        """Get the perpendicular distance from the right wall."""
-        ultrasonic = self.rightdistancesensor.distance * 100  # Convert to cm
-        laser = self.right_laser.range / 10.0  # Convert mm to cm
-        return self._perpendicular_distance(front_dist=laser,
-                                             back_dist=ultrasonic)
-
-    def _wall_angle(self,front_dist, back_dist, \
-                        sensor_gap= DISTANCE_SENSOR_DISTANCE):
-        """
-        Computes angle of wall relative to robot.
-        front_dist = distance from front sensor (cm)
-        back_dist  = distance from back sensor (cm)
-        sensor_gap = distance between sensors (cm)
-        """
-        theta_rad = math.atan2(front_dist - back_dist, sensor_gap)
-        theta_deg = math.degrees(theta_rad)
-        return theta_deg
-
-
-    def _perpendicular_distance(self,front_dist, back_dist):
-        """
-        Returns average distance as approximation of perpendicular distance.
-        (Works best if wall is roughly straight)
-        """
-        return (front_dist + back_dist) / 2.0
-
-    def get_front_distance(self) -> float:
-        """Get the distance from the front distance sensor."""
-        return self.front_distance_sensor.distance * 100  # Convert to cm
-
-    def get_front_distance_max(self) -> float:
-        """Get the maximum distance for the front distance sensor."""
-        # Check if the front distance sensor is initialized
-        return self.FRONT_DISTANCE_MAX_DISTANCE * 100  # Convert to cm
-
-    def shutdown(self) -> None:
-        try:
-            self.flush_pending_messages()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error flushing OLED messages during shutdown: %s", e)
-        if self.camera is not None:
-            self.camera.close()
-        try:
-            self.buzzer.off()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error turning off buzzer during shutdown: %s", e)
-        try:
-            self.led1.color = (0, 0, 0)
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error turning off LED1 during shutdown: %s", e)
-        try:
-            self.led1.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing LED1 during shutdown: %s", e)
-        try:
-            self.buzzer.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing buzzer during shutdown: %s", e)
-        try:
-            self.rightdistancesensor.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing right distance sensor during shutdown: %s", e)
-        try:
-            self.leftdistancesensor.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing left distance sensor during shutdown: %s", e)
-        try:
-            self.front_distance_sensor.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing front distance sensor during shutdown: %s", e)
-        try:
-            self.jumper_pin.close()
-        except Exception as e:  # pylint: disable=broad-except
-            logger.error("Error closing jumper pin during shutdown: %s", e)
-
-    def display_message(self, message: str, forceflush: bool = False) -> None:
-        """
-        Display a message on the OLED screen.
-
-        Only the last 5 messages are shown on the display.
-        """
-        now = time.time()
-        self.messages.append(message)
-        self.messages = self.messages[-5:]  # Keep only the last 5 messages
-
-        if forceflush or (now - self._last_oled_update >= self.SCREEN_UPDATE_INTERVAL):
-            # Only update the OLED if enough time has passed since the last update
-            self.flush_pending_messages()
-            self.pendingmessage = False
-
-    def flush_pending_messages(self) -> None:
-        """Flush the pending messages to the OLED display."""
-        now = time.time()
-        if self.display_loglines:
-            self.draw.rectangle((0, 0, self.SCREEN_WIDTH, self.SCREEN_HEIGHT),
-                                outline=0, fill=0)
-            for i, msg in enumerate(self.messages):
-                # Use constant for line height spacing
-                self.draw.text((0, i * RpiInterface.LINE_HEIGHT), msg, font=self.font, fill=255)
-            self.oled.image(self.image)
-            self.oled.show()
-        self._last_oled_update = now
-
-    def paint_display(self, img: Image.Image) -> None:
-        """Paint the display with the given image."""
-        self.oled.image(img)
-        self.oled.show()
-
-    def clear_messages(self) -> None:
-        """Clear all messages from the display."""
-        self.messages.clear()
-    def force_flush_messages(self) -> None:
-        """Force the OLED to update immediately."""
-        if self.pendingmessage:
-            self.flush_pending_messages()
-            self.pendingmessage = False
-
-    def get_jumper_state(self) -> bool:
-        """Get the state of the jumper pin."""
-        return self.jumper_pin.is_active
-
-    def get_acceleration(self) -> Tuple[float, float, float]:
-        """Get the acceleration from the MPU6050 sensor."""
-        accel = self.mpu.acceleration
-        return accel
-    def get_gyro(self) -> Tuple[float, float, float]:
-        """Get the gyroscope data from the MPU6050 sensor."""
-        gyro = self.mpu.gyro
-        return gyro
