+++ wroprg/src/tests/calibration.py
@@ -1,39 +1,38 @@
-
+import time
 from hardware.orientation import OrientationEstimator
-from hardware.rpi_interface import RpiInterface
+from hardware.hardware_interface import HardwareInterface
 
 
 class Calibration:
-
-
-    RpiInterface = RpiInterface(stabilize=False)
+    def __init__(self) -> None:
+        self.hw = HardwareInterface(stabilize=False)
 
     def get_acceleration(self):
         """Get the current acceleration from the RPi interface."""
-        return self.RpiInterface.get_acceleration()
+        return self.hw.get_acceleration()
 
     def get_gyro(self):
         """Get the current gyroscope data from the RPi interface."""
-        return self.RpiInterface.get_gyro()
+        return self.hw.get_gyro()
 
     def get_magnetometer(self):
         """Get the current magnetometer data from the RPi interface."""
-        return self.RpiInterface.get_magnetometer()
+        return self.hw.get_magnetometer()
 
-    def run_calibration(self):
-
+    def run_calibration(self) -> None:
         orientation_estimator = OrientationEstimator(
-                get_accel=self.get_acceleration,
-                get_gyro=self.get_gyro,
-                get_mag=self.get_magnetometer
+            get_accel=self.get_acceleration,
+            get_gyro=self.get_gyro,
+            get_mag=self.get_magnetometer,
         )
+        # Simple run to verify readings are wired; no long calibration here.
+        orientation_estimator.start_readings()
+        time.sleep(0.1)
+        orientation_estimator.shutdown()
 
-        orientation_estimator.calibrate_sensors()
 
-
-def main():
-    calibration = Calibration()
-    calibration.run_calibration()
+def main() -> None:
+    Calibration().run_calibration()
 
 
 if __name__ == "__main__":
